glamor-composite: Optimize the computation of composite region.
Computing the composite region at the composite_with_shader is very
inefficient. As when we call to here from the glamor_glyph's temproary
picture, we don't need to compute this region at all. So we move this
computing out from this function and do that at the glamor_composite
function. This can get about 5% performance gain for aa10text/rgb10text.

Signed-off-by: Zhigang Gong <zhigang.gong@linux.intel.com>