glamor-pixmap-upload: Create a uploading fbo with a texture only.
Just an initial implementation and disabled by default.
When uploading a pixmap to a texture, we don't really want
to attach the texture to any fbo. So add one fbo type
which doesn't has a gl FBO attached to it.
This commit can increase the cairo-trace's performance by
10-20%. Now the firefox-planet-gnome is 8.3s. SNA is still
the best, only take 3.5s.

Thanks for Chris to point out the A1 pixmap uploading bug.

Signed-off-by: Zhigang Gong <zhigang.gong@linux.intel.com>