polylines: Handle diagonal lines
Split a diagonal line into 2 or more horizontal/vertical lines.

We want to split the line into as small a number of segments as possible.

E.g. line from (x,y) of (1,1)->(5,2) with a slope of .25 would be split into:
(1,1)->(2,1), (3,2)->(5,2)

This is basically an implementation of Bresenham's line algorithm but with
FP instead of integers.

If the line's horizontal-ish, then iterate over the x values, and
every time the rounded y value changes, start a new rectangle.
If abs(slope) is > 1, then iterate over the y values instead.
If the slope is == 1, then we're basically stuck drawing a bunch of points.

Partially Fixes: https://bugs.freedesktop.org/show_bug.cgi?id=68524

Signed-off-by: Aaron Watry <awatry@gmail.com>
Reviewed-by: Zhigang Gong <zhigang.gong@linux.intel.com>