glamor: Add in support for the stride parameter when uploading texture data
The method __glamor_upload_pixmap_to_texture was updated to support a
stride parameter for the data being uploaded to a texture.  This
required correctly setting the alignment from 4 to a value based on the
depth of the data and also required setting GL_UNPACK_ROW_LENGTH based
on both the stride and the alignment.

The stride parameter was also updated in glamor_put_image to be
correctly specified, the old values would cause the xserver to crash.

Part of bug:
https://bugs.freedesktop.org/show_bug.cgi?id=71813

Signed-off-by: Anthony Waters <awaters1@gmail.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>