glamor: Allow nested mapping of pixmaps.
The common pattern is to do nested if statements making calls to
prepare_access() and then popping those mappings back off in each set
of braces. Some cases checked for src == dst to avoid leaking
mapping, but others didn't. Others didn't even do the nested
mappings, so a failure in the outer would result in trying to umap the
inner and failing.

By allowing nested mappings, we can fix both problems by not requiring
the care from the caller, plus we can allow a simpler nesting of all
the prepares in one if statement.

Ported from Eric's xserver glamor tree.

Signed-off-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>